send mail thanh toan

+++ src/main/java/com/tl/hotelproject/dtos/bills/VnpayCreatePaymentDto.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class VnpayCreatePaymentDto {
-    private double amount;
+    private int amount;
     private String orderId;
     private String orderInfo;
